Split a heavy test into two files (to avoid timeouts)
